fix: change receiver_memo and milestone index types from i128 to u32

+++ contracts/escrow/src/storage/types.rs
@@ -10,7 +10,7 @@ pub struct Escrow {
     pub platform_fee: u32,
     pub milestones: Vec<Milestone>,
     pub trustline: Trustline,
-    pub receiver_memo: i128,
+    pub receiver_memo: u32,
 }
 
 #[contracttype]
@@ -27,7 +27,7 @@ pub struct Milestone {
 #[contracttype]
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct MilestoneUpdate {
-    pub index: i128,
+    pub index: u32,
     pub status: String,
     pub evidence: Option<String>,
 }
